% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{classify_rpn}
\alias{classify_rpn}
\title{Classify RPN Value into Risk Level}
\usage{
classify_rpn(rpn, bands = risk_config_default()$rpn_bands)
}
\arguments{
\item{rpn}{Numeric RPN score (1--125).}

\item{bands}{Named list of band boundaries from \code{\link[=risk_config_default]{risk_config_default()}}.}
}
\value{
Character risk level name.
}
\description{
Classify RPN Value into Risk Level
}
\examples{
classify_rpn(90)
classify_rpn(25)
classify_rpn(5)

}
