% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_models.R
\name{print.risk_model}
\alias{print.risk_model}
\title{Print a Risk Model}
\usage{
\method{print}{risk_model}(x, ...)
}
\arguments{
\item{x}{A \code{risk_model} object.}

\item{...}{Ignored.}
}
\description{
Print a Risk Model
}
