% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_monte_carlo.R
\name{print.risk_monte_carlo}
\alias{print.risk_monte_carlo}
\title{Print Monte Carlo Risk Result}
\usage{
\method{print}{risk_monte_carlo}(x, ...)
}
\arguments{
\item{x}{A \code{risk_monte_carlo} object.}

\item{...}{Ignored.}
}
\description{
Print Monte Carlo Risk Result
}
