% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_reduction.R
\name{print.risk_reduction}
\alias{print.risk_reduction}
\title{Print a Risk Reduction Summary}
\usage{
\method{print}{risk_reduction}(x, ...)
}
\arguments{
\item{x}{A \code{risk_reduction} object.}

\item{...}{Ignored.}
}
\description{
Print a Risk Reduction Summary
}
