% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_register.R
\name{print.risk_register}
\alias{print.risk_register}
\title{Print Risk Register}
\usage{
\method{print}{risk_register}(x, ...)
}
\arguments{
\item{x}{A \code{risk_register} object.}

\item{...}{Ignored.}
}
\description{
Print Risk Register
}
