% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_scores.R
\name{print.risk_scores}
\alias{print.risk_scores}
\title{Print Risk Scores}
\usage{
\method{print}{risk_scores}(x, ...)
}
\arguments{
\item{x}{A \code{risk_scores} object.}

\item{...}{Ignored.}
}
\description{
Print Risk Scores
}
